+++ govuk_template_base/__init__.py
@@ -1,4 +1,4 @@
-VERSION = (0, 8)
+VERSION = (0, 9)
 __version__ = '.'.join(map(str, VERSION))
 __author__ = 'Ministry of Justice'
 __email__ = 'dev@digital.justice.gov.uk'